# Speed up min/max agg