I hereby agree to the terms of the CLA available at: https://databend.rs/policies/cla/

## Summary

Summary about this PR

```
:) select max(number) from numbers_mt(10000000000);
+-------------+
| max(number) |
+-------------+
|  9999999999 |
+-------------+
1 row in set (2.87 sec)
Read 10000000000 rows, 80 GB in 2.869 sec., 3.49 billion rows/sec., 27.89 GB/sec.

After: 
:) select max(number) from numbers_mt(10000000000);
+-------------+
| max(number) |
+-------------+
|  9999999999 |
+-------------+
1 row in set (0.54 sec)
Read 10000000000 rows, 80 GB in 0.529 sec., 18.92 billion rows/sec., 151.33 GB/sec.
```

## Changelog

 
- Improvement
 

## Related Issues

Fixes #issue

## Test Plan

Unit Tests

Stateless Tests